// Users/atishay/Documents/gpts/3j/workspace/src/lib/types.ts
export type Notification = {
  id: string;
  userId: string;
  title: string;
  description: string;
  date: string;
  read: boolean;
};